Added new Activity for AR Camera. Optimized Widget displaying method.
