+++ app/src/main/AndroidManifest.xml
@@ -7,8 +7,6 @@
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
-
-
     <uses-feature
         android:name="android.hardware.camera.ar"
         android:required="true" />
@@ -24,7 +22,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SynClient">
         <activity
-            android:name=".ARActivity"
+            android:name=".ui.ar.ARCameraActivity"
             android:exported="true" />
         <activity
             android:name=".MainActivity"
