+++ app/src/main/java/com/example/synclient/MainActivity.kt
@@ -9,8 +9,6 @@ import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.example.synclient.databinding.ActivityMainBinding
-import android.view.View
-
 
 
 class MainActivity : AppCompatActivity() {
@@ -36,4 +34,6 @@ class MainActivity : AppCompatActivity() {
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
     }
+
+
 }
