--- app/src/main/java/com/example/synclient/ui/ar/ARViewModel.kt
@@ -1,4 +1,0 @@
-package com.example.synclient.ui.ar
-
-class ARViewModel {
-}
