+++ app/src/main/res/layout/activity_ar_camera.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.ar.ARFragment">
+    tools:context=".ui.ar.ARCameraActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/scene_form_fragment"
@@ -22,32 +22,17 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0">
 
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
-
-            <ImageView
-                android:id="@+id/bear"
-                android:layout_width="100dp"
-                android:layout_height="wrap_content"
-                android:src="@drawable/ic_launcher_foreground"></ImageView>
-
-            <ImageView
-                android:id="@+id/cat"
-                android:layout_width="100dp"
-                android:layout_height="wrap_content"
-                android:src="@drawable/ic_launcher_foreground"></ImageView>
-
-            <ImageView
-                android:id="@+id/cow"
-                android:layout_width="100dp"
-                android:layout_height="wrap_content"
-                android:src="@drawable/ic_launcher_foreground"></ImageView>
-
-
-        </LinearLayout>
     </HorizontalScrollView>
+
+    <Button
+        android:id="@+id/buttonReturn"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:onClick="finishActivity"
+        android:text="Вернуться"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
